centos7: use vault.centos.org
The centos7 images try to use mirrorlist.centos.org to determine the
fastest mirror, and then download binary packages from there. But
mirrorlist.centos.org isn't online anymore. Use the vault instead.

Without this patch, the mirrors that can't be reached are ignored, and
later on it fails to properly install packages because of missing
dependencies.

Related: https://serverfault.com/a/1161847
Change-Id: I0a8f93bb16dd9245760b8a1ad53f0642d446b4cf