ttcn3-sgsn: Use 2nd IP address on emulated HNBGW
In 3G, both HNBGW and GGSN talk GTPv1U, and osmo-sgsn uses
Direct Tunnel feature to connect both together.
Since GTPv1U uses a fixed port, we need 2 IP addresses, one set for the
emulated GGSN and one for the emulated HNBGW.

This is a preparation towards adding tests validating creation of PDP
context in 3G, which was not yet tested.

Related: SYS#5435
Change-Id: I21882dc5a9f5167ac9104d4ac48d5e5631505463