debian-bookworm-titan: chown 1000 for deps
Change the ownership of the "deps" directory to a not yet created user
with the ID 1000. This user gets created when the gerrit verifications
use the container for building. Changing the ownership of the deps
directory allows using them and updating them as needed (e.g. if the
patch in review changes a commit of one of the deps).

Chowning here once results in faster jobs, compared to if we did it
every time the gerrit verifications job ran.

Creating the user here would result in conflicts with other containers
depending on debian-bookworm-titan that may create their own user with
the same UID.

Related: osmo-ci I4bea336097e8180b19b3afa41c0b984d93c21bd1
Change-Id: Id51885d0929f4966b1a4a181b3453c06d7956b53