osmo-gsm-tester: use debian-bookworm-build
osmo-gsm-tester was the last user of debian-buster-jenkins, and it just
broke again for some reason. Instead of adjusting debian-buster-jenkins
again this time, I've adjusted osmo-gsm-tester to run with
debian-bookworm-build instead so we can avoid the maintenance effort of
keeping debian-buster-jenkins working, as well as the additional time
needed to build this container.

We held off from upgrading to debian-bookworm-build, because no mongodb
debian packages are available for bookworm. However in the meantime we
have just installed the bullseye mongodb package in bookworm together
with the older libssl from bullseye that it was build against - we did
this in other containers already, so do this here too.

Other adjustments:
* Pip needs --break-system-packages now (it is fine in a docker
  container)
* Patchelf in bookworm is >= 0.11 (getting installed in the apt install
  call already), so we don't need to build it from source anymore.
* Remove 'if [ "$(arch)" = "x86_64" ];' around the mongodb code. This
  was needed earlier when we used to build this container on arm devices
  too, but AFAIK we don't do that anymore and it wasn't really useful in
  the first place. (And if we do we can bring this back easily.)
* Add rpm2cpio and cpio, because these are used during the
  osmo-gsm-tester_build-osmocom-bb job which will use this docker image
  with an upcoming osmo-ci patch (currently it is not using docker to
  build the Osmocom programs, but that leads to missing library errors
  once we upgrade to bookworm here). IMHO the build process should be
  reworked so that this is not necessary, but given that osmo-gsm-tester
  isn't really maintained currently, this is the bare minimum needed to
  migrate it away from debian buster.

Related: OS#6126
Change-Id: I5ed3b92f07f23c96b8f953e0a93991cc89476a12