debian-bookworm-build: remove pysispm, pydbus
Remove pysispm as it currently breaks building the container and as it
is unlikely that we actually need this library as it is for switching
on/off power strips.

I've seen it mentioned first in c9fa2a ("Introduce osmo-gsm-tester
docker setup") from where it apparently was copy pasted to debian 10, 11
and 12 containers afterwards c1f302c2 ("Add debian-buster-jenkins docker
setup").

Remove pydbus as well, it was apparently copy pasted with the same
history and isn't used in Osmocom software.

Fix for:
  #17 30.93 ERROR: Could not find a version that satisfies the requirement pysispm (from versions: none)
  #17 30.93 ERROR: No matching distribution found for pysispm

Change-Id: I731e171c85d530984279dc05544b6f02cec7cb3c