Fix remaining users of deprecated git.osmocom.org
git.osmocom.org is deprecated and will be removed at some point. Make
sure we don't use it anywhere in our docker containers.

Notably this drops the following line without replacement:
  ADD http://git.osmocom.org/python/osmo-python-tests/patch/?id=$OSMO_PYTHON2_TESTS_COMMIT /tmp/osmo-python-tests2-commit

This was not needed, because the patch for OSMO_PYTHON2_TESTS_COMMIT
will never change, it points to a commit hash (unlike checking master,
where the patch can change).

Change-Id: I2c7ab8fe9e87dd973df44ae1c5ff90930627040c