common/respawn: improve logging
Add "[$(date)] respawn:" infront of all log messages, so it is clear
that they come from this script and so we have timestamps that we can
compare. This will be useful when running the bpftrace scripts along
with osmo-bts-*, which will get restarted at the same time as the
osmo-bts-* processes.

Change-Id: I641f17f6988eced12c9b301042fb5c5cfd757511